requirements.txt
# Max Headroom v3.0.0 Requirements
opencv-python>=4.5.0
numpy>=1.20.0
mediapipe>=0.8.0
websocket-client>=1.0.0
websockets>=10.0.0
obs-websocket-py>=1.0.0
Pillow>=8.0.0